rename keyblock details store to the correct name
